graphql/requirements.txt
aniso8601==3.0.2
astroid==1.6.5
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.3
colormath==2.1.0
configparser==3.5.0
decorator==4.0.6
enum34==1.1.6
ExifRead==2.1.2
funcsigs==0.4
futures==3.2.0
GoogleAppEngineCloudStorageClient==1.9.22.1
graphene==2.1.3
graphene-gae==2.0.0
graphql-core==2.1
graphql-relay==0.4.5
httplib2==0.9.2
inflect==0.2.5
iso8601==0.1.12
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
networkx==1.9.1
numpy==1.9.2
pep8==1.7.1
Pillow==3.3.1
promise==2.2.1
pylint==1.9.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.10.0
typing==3.6.6
Unidecode==1.0.22
waitress==1.1.0
WebOb==1.8.2
wrapt==1.10.11
